[Triton] DSV4 fusions phase 1 (#704)
* dsv4 fusions

* update

* get rid of bf16 to fp4 upcast

* remove host side sync per decode step

* ruff clean up

* black

* clean

* rms_quant fix

* remove single stream hack

* fused_clamp_act_mul post MOE

* move kv_norm into q_norm fusion

* clean

* fix

* dummy

+++ atom/model_ops/v4_kernels/fused_compress.py
@@ -207,13 +207,13 @@ def _fused_compress_attn_kernel(
             mask=d_mask,
             other=0.0,
             eviction_policy="evict_first",
-        )
+        ).to(tl.float32)
         score_a = tl.load(
             score_in_ptr + in_row * score_in_row_stride + col_off + d,
             mask=d_mask,
             other=0.0,
             eviction_policy="evict_first",
-        )
+        ).to(tl.float32)
         ape_v = tl.load(
             ape_ptr + ape_row * dim_full + col_off + d,
             mask=d_mask,
